Flytt measure extension til egen fil
