+++ src/main/kotlin/no/nav/melosysskattehendelser/prosessering/JobMonitor.kt
@@ -159,17 +159,3 @@ class JobMonitor<T : JobMonitor.Stats>(
         fun asMap(): Map<String, Any>
     }
 }
-
-inline fun <T> measure(
-    statsMap: MutableMap<String, AtomicLong>,
-    methodName: String,
-    block: () -> T
-): T {
-    val start = System.nanoTime()
-    try {
-        return block()
-    } finally {
-        val duration = System.nanoTime() - start
-        statsMap.computeIfAbsent(methodName) { AtomicLong(0) }.addAndGet(duration)
-    }
-}
